Rename file, move helper function to another file
