+++ TestsCpp98/test/mdc/std/thread_tests.hpp
@@ -27,15 +27,15 @@
  *  to convey the resulting work.
  */
 
-#ifndef MDC_TESTS_CPP98_STD_THREADS_TESTS_HPP_
-#define MDC_TESTS_CPP98_STD_THREADS_TESTS_HPP_
+#ifndef MDC_TESTS_CPP98_STD_THREAD_TESTS_HPP_
+#define MDC_TESTS_CPP98_STD_THREAD_TESTS_HPP_
 
 namespace mdc_test {
 namespace std_test {
 
-void Threads_RunTests();
+void Thread_RunTests();
 
 } // namespace std_test
 } // namespace mdc_test
 
-#endif /* MDC_TESTS_CPP98_STD_THREADS_TESTS_HPP_ */
+#endif /* MDC_TESTS_CPP98_STD_THREAD_TESTS_HPP_ */
